Generate API key for new LDAP user
* Move backend into same module as Shibboleth backend, and rename
* Make Shib and LDAP backends both use generate_api_key helper
* Make ldap-config settings reference renamed file
